+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Changed
+### Added
+### Fixed
+
+## [5.2.0] - 2024-04-20
+### Changed
 - Cloud size monitor now runs server side.
 
 ### Added